These changes fix a couple of function arguments and reverse Dan's edits to readme.md
In function DREaD the arguments slope, freq and amp are not all needed.  The demo script was failing for lack of 'slope'.

These edits make slope, freq and amp optional, but check if the required arguments are present for either sine or linear climate variation.

And the argument new.max was missplelled in rescaleEnviro.  Fixed.